LPD-32478 Add the upgrade prefix since it is only intended to be used for upgrades. Move it to a better location, similar to the UpgradeRecorder but in kernel
